size_t to uint64 for cross platform

+++ host/src/network.c
@@ -187,7 +187,8 @@ network *make_network(int n)
     network *net = calloc(1, sizeof(network));
     net->n = n;
     net->layers = calloc(net->n, sizeof(layer));
-    net->seen = calloc(1, sizeof(size_t));
+    //net->seen = calloc(1, sizeof(size_t));
+    net->seen = calloc(1, sizeof(uint64_t));
     net->t    = calloc(1, sizeof(int));
     net->cost = calloc(1, sizeof(float));
     return net;
